% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigPLS-package.R
\docType{package}
\name{bigPLS-package}
\alias{bigPLS-package}
\title{bigPLS-package}
\description{
Provides Partial least squares Regression for regular, generalized linear and Cox models for big data. It allows for missing data in the explanatory variables. Repeated k-fold cross-validation of such models using various criteria. Bootstrap confidence intervals constructions are also available.
}
\examples{
set.seed(314)
library(bigPLS)
data(sim_data)
head(sim_data)

}
\references{
TODO
}
