% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigPLS-package.R
\docType{package}
\name{bigPLS-package}
\alias{bigPLS-package}
\alias{bigPLS-package}
\alias{bigPLS}
\title{bigPLS-package}
\description{
Provides Partial least squares Regression for regular, generalized linear and Cox models for big data. It allows for missing data in the explanatory variables. Repeated k-fold cross-validation of such models using various criteria. Bootstrap confidence intervals constructions are also available.
}
\examples{
set.seed(314)
library(bigPLS)
data(sim_data)
head(sim_data)

}
\references{
TODO
}
\seealso{
\code{\link[=big_pls_cox]{big_pls_cox()}} and \code{\link[=big_pls_cox_gd]{big_pls_cox_gd()}}
}
\author{
\strong{Maintainer}: Frederic Bertrand \email{frederic.bertrand@lecnam.net} (\href{https://orcid.org/0000-0002-0837-8281}{ORCID})

Authors:
\itemize{
  \item Myriam Maumy-Bertrand \email{myriam.maumy@ehesp.fr} (\href{https://orcid.org/0000-0002-4615-1512}{ORCID})
}

}
