% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigscale.R
\name{bigscale}
\alias{bigscale}
\title{Title}
\usage{
bigscale(
  formula = Surv(time = time, status = status) ~ .,
  data,
  norm.method = "standardize",
  strata.size = 20,
  batch.size = 1,
  features.mean = NULL,
  features.sd = NULL,
  parallel.flag = FALSE,
  num.cores = NULL,
  bigmemory.flag = FALSE,
  num.rows.chunk = 1e+06,
  col.names = NULL,
  type = "short"
)
}
\arguments{
\item{formula}{}

\item{data}{}

\item{norm.method}{}

\item{strata.size}{}

\item{batch.size}{}

\item{features.mean}{}

\item{features.sd}{}

\item{parallel.flag}{}

\item{num.cores}{}

\item{bigmemory.flag}{}

\item{num.rows.chunk}{}

\item{col.names}{}

\item{type}{}
}
\value{

}
\description{
Title
}
\examples{

1+1

}
