% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigSurvSGD.na.omit.R
\name{bigSurvSGD.na.omit}
\alias{bigSurvSGD.na.omit}
\title{Adapt bigSurvSGD to use na.omit}
\usage{
bigSurvSGD.na.omit(
  formula = Surv(time = time, status = status) ~ .,
  data,
  norm.method = "standardize",
  features.mean = NULL,
  features.sd = NULL,
  opt.method = "AMSGrad",
  beta.init = NULL,
  beta.type = "averaged",
  lr.const = 0.12,
  lr.tau = 0.5,
  strata.size = 20,
  batch.size = 1,
  num.epoch = 100,
  b1 = 0.9,
  b2 = 0.99,
  eps = 1e-08,
  inference.method = "plugin",
  num.boot = 1000,
  num.epoch.boot = 100,
  boot.method = "SGD",
  lr.const.boot = 0.12,
  lr.tau.boot = 0.5,
  num.sample.strata = 1000,
  sig.level = 0.05,
  beta0 = 0,
  alpha = NULL,
  lambda = NULL,
  nlambda = 100,
  num.strata.lambda = 10,
  lambda.scale = 1,
  parallel.flag = FALSE,
  num.cores = NULL,
  bigmemory.flag = FALSE,
  num.rows.chunk = 1e+06,
  col.names = NULL,
  type = "float"
)
}
\arguments{
\item{formula}{}

\item{data}{}

\item{norm.method}{}

\item{features.mean}{}

\item{features.sd}{}

\item{opt.method}{}

\item{beta.init}{}

\item{beta.type}{}

\item{lr.const}{}

\item{lr.tau}{}

\item{strata.size}{}

\item{batch.size}{}

\item{num.epoch}{}

\item{b1}{}

\item{b2}{}

\item{eps}{}

\item{inference.method}{}

\item{num.boot}{}

\item{num.epoch.boot}{}

\item{boot.method}{}

\item{lr.const.boot}{}

\item{lr.tau.boot}{}

\item{num.sample.strata}{}

\item{sig.level}{}

\item{beta0}{}

\item{alpha}{}

\item{lambda}{}

\item{nlambda}{}

\item{num.strata.lambda}{}

\item{lambda.scale}{}

\item{parallel.flag}{}

\item{num.cores}{}

\item{bigmemory.flag}{}

\item{num.rows.chunk}{}

\item{col.names}{}

\item{type}{}
}
\value{
coef: Log of hazards ratio. If no inference is used, it returns a vector for estimated coefficients: If inference is used, it returns a matrix including estimates and confidence intervals of coefficients. In case of penalization, it resturns a matrix with columns corresponding to lambdas.
coef.exp: Exponentiated version of coef (hazards ratio).
lambda: Returns lambda(s) used for penalizarion.
alpha: Returns alpha used for penalizarion.
features.mean: Returns means of features, if given or calculated
features.sd: Returns standard deviations of features, if given or calculated.
}
\description{
Adapt bigSurvSGD to use na.omit
}
\examples{
#' set.seed(314)
library(sim_data)

}
\seealso{
See Also \code{\link[bigSurvSGD]{bigSurvSGDbigSurvSGD}}
}
