% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigSurvSGD.na.omit.R
\name{partialbigSurvSGDv0}
\alias{partialbigSurvSGDv0}
\title{Title}
\usage{
partialbigSurvSGDv0(
  name.col,
  datapath,
  ncores = 1,
  resBigscale = resultsBigscale,
  bigmemory.flag = FALSE,
  parallel.flag = FALSE,
  inf.mth = "none"
)
}
\arguments{
\item{name.col}{}

\item{datapath}{}

\item{ncores}{}

\item{resBigscale}{}

\item{bigmemory.flag}{}

\item{inf.mth}{}
}
\value{

}
\description{
Title
}
\examples{

1+1


}
